refactor(app): extract ProgressTracker into app/progress.rs
OpStep, OpProgress, and ProgressTracker move verbatim, together with
their pure unit tests (split out of the former text_edit_tests
grab-bag; the quit-during-progress tests stay in mod.rs as
quit_during_progress_tests since they drive App::handle_key).
Re-exported from app so external paths are unchanged. Part of #270.
